Use entry_points to install the gtabview executable
Supersedes #23, fixes #20 and hopefully unblocks #24.

Move gtabview into a dedicated gtabview_cli package in order to
ensure the gtabview module is loaded on demand as previously done.

This is still not ideal. The graphical initialization of gtabview should
probably be moved to a submodule of gtabview itself (such as
gtabview.view) so that models can be used/tested without QT to begin
with (see #33)