Use IPython's GUI hooks instead of a DetachedViewController
Instead of fighting with QT's main loop, use the python interpreter hook
to "fake" interactivity. Leverage IPython's "enable_gui" to run the main
QT event loop for us, which is what matplotlib is internally doing.

Check for ipython inclusion, which is always true for Jupyter. In this
case, initialize the 'qt' backend. This fixes #32 in any mode
(inline/detached mpl graphics).